refactor: news bar 모듈 분리
