+++ src/news-bar/rolling.js
@@ -1,3 +1,5 @@
+import { createNewsBar } from "./setNewsBar.js";
+
 const setRollingEvent = (rollingElement) => {
     window.setInterval(() => {
         rollingElement.style.transitionDuration = "400ms";
@@ -15,31 +17,9 @@ const infiniteRolling = () => {
     rollingElement.forEach((item) => setRollingEvent(item));
 };
 
-const getNewsBarElement = (newData) => {
-    const { press, headline } = newData;
-    return `
-        <li>
-            <span class="display-bold14 news-bar-press">${press}</span>
-            <span class="available-medium14 news-bar-headline">${headline}</span>
-        </li>
-        `;
-};
-
-const createNewsBarData = (item, index) => {
-    const $ul = document.createElement("ul");
-    fetch(`../data/news-bar-data${index + 1}.json`)
-        .then((response) => response.json())
-        .then((jsonData) => {
-            jsonData.forEach((item) =>
-                $ul.insertAdjacentHTML("beforeend", getNewsBarElement(item))
-            );
-        });
-    item.appendChild($ul);
-};
-
 const setNewsBarRolling = () => {
     const newsBarRolling = document.querySelectorAll(".news-bar-rolling");
-    newsBarRolling.forEach((item, index) => createNewsBarData(item, index));
+    newsBarRolling.forEach((item, index) => createNewsBar(item, index));
 };
 
 export { setNewsBarRolling, infiniteRolling };
